🧪 refactor(database/alembic): change bot_config.chat_key type to BigInteger 🧪 refactor(db/mongo.py): add local_test helper and clean up test code 🧪 refactor(shared/infrastructure/database/models.py): change chat_key type to BigInteger

+++ shared/infrastructure/database/models.py
@@ -145,7 +145,7 @@ class KVStore(Base):
 class BotConfig(Base):
     __tablename__ = 'bot_config'
     chat_id = Column(BigInteger, primary_key=True)
-    chat_key = Column(Integer, primary_key=True)
+    chat_key = Column(BigInteger, primary_key=True)
     chat_key_name = Column(String)
     chat_value = Column(JSONB) # Using JSONB for flexibility
